requirements.txt
hydra-core
selenium
selenium_stealth
2captcha-python
omegaconf
python-dotenv